feat: landing page after logging in (#2081)
