+++ space/components/accounts/index.ts
@@ -4,3 +4,5 @@ export * from "./email-reset-password-form";
 export * from "./github-login-button";
 export * from "./google-login";
 export * from "./onboarding-form";
+export * from "./sign-in";
+export * from "./user-logged-in";
